[Bronze V] Title: 이상한 기호, Time: 80 ms, Memory: 11556 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,12 +7,12 @@ public class Main {
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int a = Integer.parseInt(st.nextToken());
-		int b = Integer.parseInt(st.nextToken());
+		long a = Long.parseLong(st.nextToken());
+		long b = Long.parseLong(st.nextToken());
 		System.out.println(func(a, b));
 		br.close();
 	}
-	public static int func(int a, int b){
+	public static long func(long a, long b){
 		return (a+b)*(a-b);
 	}
 }
